Per-feed teams instead of per package
This change reduces security slightly because now you can write to the cache for any package in an org so long as you have write access to one of them, and you can read the cache for any package in an org so long as you can read one of them.

The second one is the main issue and we can't release this plugin before it is fixed, as it will leak private build artefacts (although not source code). The plan to fix this is to look into the uploaded cache artefacts and try to figure out what package they point to somehow, and base the permissions on that.